Isolated the entire ConnectionsList concept to its own module.
