+++ app/react/ConnectionsList/actions/actions.js
@@ -1,0 +1,82 @@
+import {actions} from 'app/BasicReducer';
+import {actions as formActions} from 'react-redux-form';
+import {notify} from 'app/Notifications';
+
+import referencesAPI from 'app/Viewer/referencesAPI';
+import {fromJS as Immutable} from 'immutable';
+import {get as prioritySortingCriteria} from 'app/utils/prioritySortingCriteria';
+
+import * as uiActions from 'app/Entities/actions/uiActions';
+
+export function searchReferences() {
+  return function (dispatch, getState) {
+    const connectionsList = getState().connectionsList;
+    const {entityId, sort, filters} = connectionsList;
+    const searchTerm = connectionsList.search && connectionsList.search.searchTerm ? connectionsList.search.searchTerm.value : '';
+    const options = filters.merge(sort).merge({searchTerm});
+
+    return referencesAPI.search(entityId, options.toJS())
+    .then(results => {
+      dispatch(actions.set('connectionsList/searchResults', results));
+      dispatch(uiActions.showTab('references'));
+    });
+  };
+}
+
+export function connectionsChanged() {
+  return function (dispatch, getState) {
+    const connectionsList = getState().connectionsList;
+    const {entityId} = connectionsList;
+
+    return referencesAPI.getGroupedByConnection(entityId)
+    .then(connectionsGroups => {
+      const filteredTemplates = connectionsGroups.reduce((templateIds, group) => {
+        return templateIds.concat(group.templates.map(t => t._id.toString()));
+      }, []);
+
+      const sortOptions = prioritySortingCriteria({currentCriteria: connectionsList.sort, filteredTemplates, templates: getState().templates});
+      return Promise.all([connectionsGroups, sortOptions]);
+    })
+    .then(([connectionsGroups, sort]) => {
+      dispatch(actions.set('connectionsList/connectionsGroups', connectionsGroups));
+      dispatch(formActions.merge('connectionsList.sort', sort));
+      return searchReferences()(dispatch, getState);
+    });
+  };
+}
+
+export function deleteConnection(connection) {
+  return function (dispatch, getState) {
+    return referencesAPI.delete(connection)
+    .then(() => {
+      dispatch(notify('Connection deleted', 'success'));
+      return connectionsChanged()(dispatch, getState);
+    });
+  };
+}
+
+export function loadMoreReferences(limit) {
+  return function (dispatch, getState) {
+    const connectionsList = getState().connectionsList;
+    dispatch(actions.set('connectionsList/filters', connectionsList.filters.set('limit', limit)));
+    return searchReferences()(dispatch, getState);
+  };
+}
+
+export function setFilter(groupFilterValues) {
+  return function (dispatch, getState) {
+    const connectionsList = getState().connectionsList;
+    const currentFilter = connectionsList.filters.get('filter') || Immutable({});
+    const newFilter = currentFilter.merge(groupFilterValues);
+    dispatch(actions.set('connectionsList/filters', connectionsList.filters.set('filter', newFilter)));
+    return searchReferences()(dispatch, getState);
+  };
+}
+
+export function resetSearch() {
+  return function (dispatch, getState) {
+    dispatch(formActions.change('connectionsList/search.searchTerm', ''));
+    dispatch(actions.set('connectionsList/filters', Immutable({})));
+    return searchReferences()(dispatch, getState);
+  };
+}
