+++ app/react/ConnectionsList/components/SearchBar.js
