+++ app/react/ConnectionsList/index.js
@@ -1,0 +1,11 @@
+import reducer from './reducers/reducer';
+import ResetSearch from './components/ResetSearch';
+import ConnectionsGroups from './components/ConnectionsGroups';
+import ConnectionsList from './components/ConnectionsList';
+
+export default {
+  reducer,
+  ResetSearch,
+  ConnectionsGroups,
+  ConnectionsList
+};
